Ensure vrrp data is not accessed in dbus thread
Keepalived is not written to be pthread thread safe. It is therefore
necessary to access all vrrp data in the main thread. This commit adds
message queues between the two threads, so that all vrrp data is
accessed in a synchronous manner in the main thread, and passed via
a queue to the dbus thread.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>