Replace dbus_queue_ent fields with single GVariant
Removed ent->str and ent->val. ent->args should be a GVariant of type
a{sv} (dictionary with string keys and variant values) when starting the
thread, and any appropriate GVariant type when returning (the only time we need
a return value is for getting properties, in which case it's a GVariant
of type (s) or (v)).